Various minor improvements, cleanup
* Pass remote type to supertype in constructor as it's needed at
object creation time
* Use a LinkedList for items to improve performance when removing items
* Use float instead of double for values to preserve memory (float is
  precise enough for charts)
* Allow font to be reset to null

+++ examples/org.eclipse.rap.demo.d3chart/src/org/eclipse/rap/demo/d3chart/internal/BarChartExample.java
@@ -57,7 +57,7 @@ public class BarChartExample implements IExamplePage {
     composite.setLayoutData( ExampleUtil.createFillData() );
     createButton( composite, "Add item", new Listener() {
       public void handleEvent( Event event ) {
-        double value = Math.random() * 0.8;
+        float value = ( float )( Math.random() * 0.8 );
         Color color = new Color( barChart.getDisplay(), colors.next() );
         addItem( pieChart, value, color );
         addItem( barChart, value, color );
@@ -110,7 +110,7 @@ public class BarChartExample implements IExamplePage {
     return button;
   }
 
-  private void addItem( Chart chart, double value, Color color ) {
+  private void addItem( Chart chart, float value, Color color ) {
     ChartItem item = new ChartItem( chart );
     item.setValue( value );
     DecimalFormat format = new DecimalFormat( "#.#" );
